refactor(deprecate): extract _dispatch engine module
- move target resolution (`_normalize_target`, cross-class + positional-only guards, `_warn_stacking_misconfiguration`) and the call-plan runtime (`_build_call_plan`, kwargs assembly, `_invoke_sync`/`_invoke_async`, `_cycle_detection`) into new `_dispatch.py`
- reduce `callables.py` to packing descriptors and the public `deprecated` / `deprecated_callable` decorators, importing the engine from `_dispatch` (one-way; no cycle)
- update the `deprecation.py` shim and `proxy` to source the relocated helpers from `_dispatch`

---
